Add databse connectivity.
This commit adds the ability to load data from a database with the
intention to use this data to create objects for the simulation.

- Add DatabaseConnection.java. Access data from the database.
- Add OrbitalReciever.java. Process data to create objects for the
simulation.

Also in OrbitalModel, adds the ability to easily switch between full and
simplified gravity in preparation for adding potentially thousands of
small astronomical objects (asteroids) into the simulation. Simplified
gravity is accurate enough for these objects and will increase
perfomance substantially.