Add ateroids obtained from the database into the simulation.
Asteroids created from database data now display in the simulation with
random values for initial position, initial velocity, size and color.

Using the factory method pattern, data extracted from the database in
DatabaseConnection.queryDatabase() is packaged into a object type
specified by the factory class.

The factory class is given as a parameter when the database is queried.